requirements.txt
aiogram>=3.15.0
Pillow>=10.0.0
Babel>=2.13.0
SQLModel>=0.0.14
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
google-genai>=0.1.0
